+++ src/FlaUILibrary/flaui/module/element.py
@@ -30,6 +30,8 @@ class Element(ModuleInterface):
         """
         GET_ELEMENT = "GET_ELEMENT"
         GET_ELEMENT_NAME = "GET_ELEMENT_NAME"
+        GET_ELEMENT_BY_XPATH = "GET_ELEMENT_BY_XPATH"
+        FOCUS_ELEMENT_BY_XPATH = "FOCUS_ELEMENT_BY_XPATH"
         GET_ELEMENT_RECTANGLE_BOUNDING = "GET_ELEMENT_RECTANGLE_BOUNDING"
         FOCUS_ELEMENT = "FOCUS_ELEMENT"
         FIND_ALL_ELEMENTS = "FIND_ALL_ELEMENTS"
@@ -88,6 +90,10 @@ class Element(ModuleInterface):
                 lambda: self._get_element(values["xpath"]).Focus(),
             self.Action.GET_ELEMENT:
                 lambda: self._get_element(values["xpath"]),
+            self.Action.GET_ELEMENT_BY_XPATH:
+                lambda: self._get_element_by_xpath(values["xpath"]),
+            self.Action.FOCUS_ELEMENT_BY_XPATH:
+                lambda: self._get_element_by_xpath(values["xpath"]).focus(),
             self.Action.GET_ELEMENT_NAME:
                 lambda: self._get_name_from_element(values["xpath"]),
             self.Action.GET_ELEMENT_RECTANGLE_BOUNDING: